requirements.txt
torch
torchvision
pandas
Pillow
spacy
pytorch-lightning
numpy
matplotlib
scikit-learn